+++ src/Model/Equal.java
@@ -5,7 +5,7 @@ import java.util.List;
 public class Equal extends Node {
 	
 	@Override
-	public int interpret() {
+	public double interpret() {
 		List<Node> children = getChildren();
 		if (children.get(0).interpret() == children.get(1).interpret()) {
 			return 1;
